Python Files/requirements.txt
streamlit
mysql-connector-python
openai
python-dotenv